+++ client/js/components/procedure/StatementSegmentsList/StatementSegment.vue
@@ -62,7 +62,7 @@
                     {{ Object.values(customFields).find(field => field.id === customField.id)?.attributes?.name || '' }}:
                   </dt>
                   <dd>
-                    {{ customField.value ? customField.value : Translator.trans('not.assigned')}}
+                    {{ customField.value ? customField.value : Translator.trans('not.assigned') }}
                   </dd>
                 </div>
               </template>
@@ -385,9 +385,11 @@
           }"
           data-cy="segmentMap"
           @click.prevent="showMap">
-          <i
-            class="fa fa-map-marker"
-            aria-hidden="true" />
+          <dp-icon
+            class="mx-auto"
+            icon="map-pin"
+            :weight="hasPolygonFeatures() ? 'fill' : 'regular'"
+          />
         </button>
       </div>
     </div>
@@ -554,7 +556,7 @@ export default {
       if (this.segment?.relationships?.assignee?.data?.id && this.segment.relationships.assignee.data.id !== '') {
         const assignee = this.assignableUserItems[this.segment.relationships.assignee.data.id]
         const name = `${assignee.attributes.firstname} ${assignee.attributes.lastname}`
-        const orga = assignee ? assignee.rel('orga') : ''
+        const orga = assignee?.rel ? assignee.rel('orga') : ''
 
         return { id: this.segment.relationships.assignee.data.id, name, orgaName: orga ? orga.attributes.name : '' }
       } else {
@@ -642,14 +644,6 @@ export default {
   },
 
   methods: {
-    ...mapActions('AssignableUser', {
-      fetchAssignableUsers: 'list'
-    }),
-
-    ...mapActions('Place', {
-      fetchPlaces: 'list'
-    }),
-
     ...mapActions('SegmentSlidebar', [
       'toggleSlidebarContent'
     ]),
@@ -717,7 +711,7 @@ export default {
         }
       }
 
-      dpApi.patch(Routing.generate('api_resource_update', { resourceType: 'StatementSegment', resourceId: this.segment.id }), {}, payload)
+      return dpApi.patch(Routing.generate('api_resource_update', { resourceType: 'StatementSegment', resourceId: this.segment.id }), {}, payload)
         .then(checkResponse)
         .then(() => {
           this.claimLoading = false
@@ -742,6 +736,19 @@ export default {
 
     handleTabChange (id) {
       this.activeId = id
+    },
+
+    hasPolygonFeatures () {
+      const raw = this.segment.attributes.polygon
+
+      if (!raw || typeof raw !== 'string') {
+        return false
+      }
+
+      const parsedPolygon = JSON.parse(raw)
+      const features = parsedPolygon?.features
+
+      return Array.isArray(features) && features.length > 0
     },
 
     initAssignableUsers () {
@@ -836,6 +843,9 @@ export default {
         data: {
           id: this.segment.id,
           type: 'StatementSegment',
+          attributes: {
+            recommendation: this.segment.attributes.recommendation
+          },
           relationships: {
             assignee,
             place
@@ -852,12 +862,14 @@ export default {
         type: 'StatementSegment',
         attributes: {
           ...this.segment.attributes,
-          ...(hasCustomFields ? {
-            customFields: {
-              ...this.segment.attributes.customFields,
-              ...payload.data.attributes?.customFields
-            }
-          } : {})
+          ...(hasCustomFields
+            ? {
+                customFields: {
+                  ...this.segment.attributes.customFields,
+                  ...payload.data.attributes?.customFields
+                }
+              }
+            : {})
         },
         relationships: {
           ...this.segment.relationships,