+++ CHANGELOG.md
@@ -4,8 +4,16 @@
 - **Minor Version**: Incremented for each release.
 - **Patch Version**: Incremented for bug fixes.
 
-
 ## UNRELEASED
+- Export Original Statements as docx in the Statement List
+- Allow filtering of institution tags in DpAddOrganizationList
+- Allow filtering of institution tags in AdminstrationMemberList / refactor twig
+
+## v4.3.0 (2025-06-13)
+- Add Versioning of custom fields of segments
+- Update Elasticsearch to version 8
+- Export Original Statements as csv in the Statement List
+
 ## v4.1.0 (2025-05-21)
 ## v4.0.0 (2025-05-21)
 - Update to symfony 6.4